requirements.txt
# Core Web Framework
Flask==3.0.0
gunicorn==21.2.0
werkzeug==3.0.1
Jinja2==3.1.3

# AI & Voice Services
groq==0.4.2
openai==1.12.0
edge-tts==6.1.9

# Database & Auth
supabase==2.3.4

# Utilities & Config
python-dotenv==1.0.0
httpx==0.27.0

# Document Processing / Generation
python-docx==1.1.0
PyPDF2==3.0.1
xhtml2pdf==0.2.13
markdown==3.5.2
